fix(mcp): finish protocol review corrections
